отображение фильтр и поиск сотрудников

+++ Models/Tours.cs
@@ -24,7 +24,7 @@ namespace TravelAgencyCRM.Models
         public string Country { get; set; }
 
         [Column(TypeName = "money")]
-        public decimal? Price { get; set; }
+        public int? Price { get; set; }
 
         public int? State { get; set; }
